chore(engine-v2): split engine.rs
The file was getting increasingly big so split it up into separate
modules for clarity.
