+++ engine/crates/engine-v2/src/engine.rs
@@ -1,45 +1,25 @@
-use ::runtime::{
-    auth::AccessToken,
-    error::ErrorResponse,
-    hooks::Hooks,
-    operation_cache::{OperationCache, OperationCacheFactory},
-    rate_limiting::RateLimitKey,
-};
-use async_runtime::stream::StreamExt as _;
+use ::runtime::{hooks::Hooks, operation_cache::OperationCacheFactory};
 use bytes::Bytes;
-use engine_parser::types::OperationType;
-use enumset::EnumSet;
-use futures::{channel::mpsc, FutureExt, StreamExt, TryFutureExt};
-use futures_util::{SinkExt, Stream};
+use futures::{StreamExt, TryFutureExt};
+use futures_util::Stream;
 use gateway_v2_auth::AuthService;
-use grafbase_telemetry::{
-    gql_response_status::GraphqlResponseStatus,
-    grafbase_client::Client,
-    metrics::{
-        GraphqlErrorAttributes, GraphqlOperationMetrics, GraphqlRequestMetricsAttributes, OperationMetricsAttributes,
-        QueryPreparationAttributes,
-    },
-    span::{gql::GqlRequestSpan, GqlRecorderSpanExt, GRAFBASE_TARGET},
-};
+use grafbase_telemetry::metrics::EngineMetrics;
 use retry_budget::RetryBudgets;
 use schema::Schema;
 use std::{borrow::Cow, future::Future, sync::Arc};
-use tracing::Instrument;
-use trusted_documents::OperationDocument;
-use web_time::{Instant, SystemTime};
 
 use crate::{
-    execution::{ExecutableOperation, PreExecutionContext},
     graphql_over_http::{Http, ResponseFormat, StreamingResponseFormat},
-    operation::{Operation, PreparedOperation, Variables},
-    request::{BatchRequest, QueryParamsRequest, Request},
-    response::{GraphqlError, RefusedRequestResponse, RequestErrorResponse, Response},
-    websocket, Body, ErrorCode,
+    operation::PreparedOperation,
+    response::Response,
+    websocket, Body,
 };
+pub(crate) use execute::RequestContext;
 use runtime::RuntimeExt;
 
 mod cache;
 mod error_responses;
+mod execute;
 mod retry_budget;
 mod runtime;
 mod trusted_documents;
@@ -62,7 +42,7 @@ pub struct Engine<R: Runtime> {
     pub(crate) schema: Arc<Schema>,
     pub(crate) schema_version: SchemaVersion,
     pub(crate) runtime: R,
-    pub(crate) operation_metrics: GraphqlOperationMetrics,
+    pub(crate) metrics: EngineMetrics,
     auth: AuthService,
     retry_budgets: RetryBudgets,
     operation_cache: <R::OperationCacheFactory as OperationCacheFactory>::Cache<Arc<PreparedOperation>>,
@@ -95,7 +75,7 @@ impl<R: Runtime> Engine<R> {
             }),
             auth,
             retry_budgets: RetryBudgets::build(&schema),
-            operation_metrics: GraphqlOperationMetrics::build(runtime.meter()),
+            metrics: EngineMetrics::build(runtime.meter()),
             operation_cache: runtime.operation_cache_factory().create().await,
             schema,
             runtime,
@@ -108,72 +88,23 @@ impl<R: Runtime> Engine<R> {
     where
         F: Future<Output = Result<Bytes, (http::StatusCode, String)>> + Send,
     {
-        // Following the recommendation of the GraphQL over HTTP spec to require a valid Accept
-        // header
-        let (parts, body) = request.into_parts();
-        let Some(format) = ResponseFormat::extract_from(&parts.headers, self.default_response_format) else {
-            // GraphQL-over-HTTP spec:
-            //   In alignment with the HTTP 1.1 Accept specification, when a client does not include at least one supported media type in the Accept HTTP header, the server MUST either:
-            //     1. Respond with a 406 Not Acceptable status code and stop processing the request (RECOMMENDED); OR
-            //     2. Disregard the Accept header and respond with the server's choice of media type (NOT RECOMMENDED).
-            return Http::from(
-                self.default_response_format,
-                RefusedRequestResponse::not_acceptable_error(),
-            );
+        let (
+            http::request::Parts {
+                method, headers, uri, ..
+            },
+            body,
+            response_format,
+        ) = match self.unpack_http_request(request) {
+            Ok(req) => req,
+            Err(response) => return response,
         };
-
-        if parts.method == http::Method::POST {
-            // GraphQL-over-HTTP spec:
-            //   If the client does not supply a Content-Type header with a POST request,
-            //   the server SHOULD reject the request using the appropriate 4xx status code.
-            if !content_type_is_application_json(&parts.headers) {
-                return Http::from(format, RefusedRequestResponse::unsupported_media_type());
-            }
-        } else if parts.method != http::Method::GET {
-            return Http::from(
-                format,
-                RefusedRequestResponse::method_not_allowed("Only GET or POST are supported."),
-            );
-        }
 
         let request_context_fut = self
-            .create_request_context(!parts.method.is_safe(), parts.headers, format)
-            .map_err(|response| Http::from(format, response));
+            .create_request_context(!method.is_safe(), headers, response_format)
+            .map_err(|response| Http::from(response_format, response));
 
-        let graphql_request_fut = async move {
-            if parts.method == http::Method::POST {
-                let body = body.await.map_err(|(status, message)| {
-                    Http::from(
-                        format,
-                        Response::refuse_request_with(status, GraphqlError::new(message, ErrorCode::BadRequest)),
-                    )
-                })?;
-
-                self.operation_metrics.record_request_body_size(body.len());
-
-                serde_json::from_slice(&body).map_err(|err| {
-                    Http::from(
-                        format,
-                        RefusedRequestResponse::not_well_formed_graphql_over_http_request(format_args!(
-                            "JSON deserialization failure: {err}",
-                        )),
-                    )
-                })
-            } else {
-                let query = parts.uri.query().unwrap_or_default();
-
-                serde_urlencoded::from_str::<QueryParamsRequest>(query)
-                    .map(|request| BatchRequest::Single(request.into()))
-                    .map_err(|err| {
-                        Http::from(
-                            format,
-                            RefusedRequestResponse::not_well_formed_graphql_over_http_request(format_args!(
-                                "Could not deserialize request from query parameters: {err}"
-                            )),
-                        )
-                    })
-            }
-        };
+        let graphql_request_fut =
+            self.extract_well_formed_graphql_over_http_request(method, uri, response_format, body);
 
         // Retrieve the request body while processing the headers
         match futures::try_join!(request_context_fut, graphql_request_fut) {
@@ -204,428 +135,6 @@ impl<R: Runtime> Engine<R> {
             request_context: Arc::new(request_context),
         })
     }
-
-    async fn create_request_context(
-        &self,
-        mutations_allowed: bool,
-        headers: http::HeaderMap,
-        response_format: ResponseFormat,
-    ) -> Result<RequestContext<<R::Hooks as Hooks>::Context>, Response> {
-        let client = Client::extract_from(&headers);
-
-        let (hooks_context, headers) = self
-            .runtime
-            .hooks()
-            .on_gateway_request(headers)
-            .await
-            .map_err(|ErrorResponse { status, error }| Response::refuse_request_with(status, error))?;
-
-        let Some(access_token) = self.auth.authenticate(&headers).await else {
-            return Err(RefusedRequestResponse::unauthenticated());
-        };
-
-        // Currently it doesn't rely on authentication, but likely will at some point.
-        if self.runtime.rate_limiter().limit(&RateLimitKey::Global).await.is_err() {
-            return Err(RefusedRequestResponse::gateway_rate_limited());
-        }
-
-        Ok(RequestContext {
-            mutations_allowed,
-            headers,
-            response_format,
-            client,
-            access_token,
-            hooks_context,
-        })
-    }
-
-    async fn execute_well_formed_graphql_request(
-        self: &Arc<Self>,
-        request_context: RequestContext<<R::Hooks as Hooks>::Context>,
-        request: BatchRequest,
-    ) -> http::Response<Body> {
-        match request {
-            BatchRequest::Single(request) => {
-                if let ResponseFormat::Streaming(format) = request_context.response_format {
-                    Http::stream(format, self.execute_stream(Arc::new(request_context), request)).await
-                } else {
-                    let Some(response) = self
-                        .runtime
-                        .with_timeout(
-                            self.schema.settings.timeout,
-                            self.execute_single(&request_context, request),
-                        )
-                        .await
-                    else {
-                        return Http::from(request_context.response_format, RequestErrorResponse::gateway_timeout());
-                    };
-                    Http::from(request_context.response_format, response)
-                }
-            }
-            BatchRequest::Batch(requests) => {
-                let ResponseFormat::Complete(format) = request_context.response_format else {
-                    return Http::from(
-                        request_context.response_format,
-                        RequestErrorResponse::bad_request_but_well_formed_graphql_over_http_request(
-                            "batch requests cannot be returned as multipart or event-stream responses",
-                        ),
-                    );
-                };
-
-                self.operation_metrics.record_batch_size(requests.len());
-
-                let Some(responses) = self
-                    .runtime
-                    .with_timeout(
-                        self.schema.settings.timeout,
-                        futures_util::stream::iter(requests.into_iter())
-                            .then(|request| self.execute_single(&request_context, request))
-                            .collect::<Vec<_>>(),
-                    )
-                    .await
-                else {
-                    return Http::from(request_context.response_format, RequestErrorResponse::gateway_timeout());
-                };
-
-                Http::batch(format, responses)
-            }
-        }
-    }
-
-    async fn execute_single(
-        &self,
-        request_context: &RequestContext<<R::Hooks as Hooks>::Context>,
-        request: Request,
-    ) -> Response {
-        let start = Instant::now();
-        let span = GqlRequestSpan::create();
-
-        async {
-            let ctx = PreExecutionContext::new(self, request_context);
-            let (operation_metrics_attributes, response) = ctx.execute_single(request).await;
-
-            let elapsed = start.elapsed();
-            let status = response.graphql_status();
-
-            if let Some(operation_metrics_attributes) = operation_metrics_attributes.clone() {
-                span.record_gql_request((&operation_metrics_attributes).into());
-
-                self.operation_metrics.record_operation_duration(
-                    GraphqlRequestMetricsAttributes {
-                        operation: operation_metrics_attributes,
-                        status,
-                        cache_status: None,
-                        client: request_context.client.clone(),
-                    },
-                    elapsed,
-                );
-            }
-
-            span.record_gql_status(status);
-
-            if status.is_success() {
-                tracing::debug!(target: GRAFBASE_TARGET, "gateway request")
-            } else {
-                let message = response
-                    .errors()
-                    .first()
-                    .map(|error| error.message.clone())
-                    .unwrap_or_else(|| String::from("gateway error").into());
-
-                tracing::debug!(target: GRAFBASE_TARGET, "{message}")
-            }
-
-            if let Some(attributes) = operation_metrics_attributes {
-                let errors = response
-                    .errors()
-                    .iter()
-                    .fold(EnumSet::<ErrorCode>::empty(), |mut set, error| {
-                        set |= error.code;
-                        set
-                    });
-
-                for error in errors {
-                    self.operation_metrics.increment_graphql_errors(GraphqlErrorAttributes {
-                        code: error.into(),
-                        operation_name: attributes.name.clone(),
-                        client: request_context.client.clone(),
-                    });
-                }
-            }
-
-            response
-        }
-        .instrument(span.clone())
-        .await
-    }
-
-    fn execute_stream(
-        self: &Arc<Self>,
-        request_context: Arc<RequestContext<<R::Hooks as Hooks>::Context>>,
-        request: Request,
-    ) -> impl Stream<Item = Response> + Send + 'static {
-        let start = Instant::now();
-        let engine = Arc::clone(self);
-        let (sender, receiver) = mpsc::channel(2);
-
-        let span = GqlRequestSpan::create();
-        let span_clone = span.clone();
-
-        receiver.join(
-            async move {
-                let ctx = PreExecutionContext::new(&engine, &request_context);
-                let (operation_metrics_attributes, status) = ctx.execute_stream(request, sender).await;
-                let elapsed = start.elapsed();
-
-                if let Some(operation_metrics_attributes) = operation_metrics_attributes {
-                    tracing::Span::current().record_gql_request((&operation_metrics_attributes).into());
-
-                    engine.operation_metrics.record_operation_duration(
-                        GraphqlRequestMetricsAttributes {
-                            operation: operation_metrics_attributes,
-                            status,
-                            cache_status: None,
-                            client: request_context.client.clone(),
-                        },
-                        elapsed,
-                    );
-                }
-
-                span.record_gql_status(status);
-
-                if status.is_success() {
-                    tracing::debug!(target: GRAFBASE_TARGET, "gateway request")
-                } else {
-                    tracing::debug!(target: GRAFBASE_TARGET, "gateway error")
-                }
-            }
-            .instrument(span_clone),
-        )
-    }
-}
-
-impl<'ctx, R: Runtime> PreExecutionContext<'ctx, R> {
-    async fn execute_single(mut self, request: Request) -> (Option<OperationMetricsAttributes>, Response) {
-        let operation_plan = match self.prepare_operation(request).await {
-            Ok(operation_plan) => operation_plan,
-            Err((metadata, response)) => return (metadata, response),
-        };
-
-        let metrics_attributes = Some(operation_plan.metrics_attributes.clone());
-        let response = if matches!(operation_plan.ty(), OperationType::Subscription) {
-            RequestErrorResponse::bad_request_but_well_formed_graphql_over_http_request(
-                "Subscriptions are only suported on streaming transports. Try making a request with SSE or WebSockets",
-            )
-        } else {
-            self.execute_query_or_mutation(operation_plan).await
-        };
-
-        (metrics_attributes, response)
-    }
-
-    async fn execute_stream(
-        mut self,
-        request: Request,
-        mut sender: mpsc::Sender<Response>,
-    ) -> (Option<OperationMetricsAttributes>, GraphqlResponseStatus) {
-        let engine = self.engine;
-        let client = self.request_context.client.clone();
-
-        // If it's a subscription, we at least have a timeout on the operation preparation.
-        let result = engine
-            .runtime
-            .with_timeout(self.engine.schema.settings.timeout, async {
-                let operation_plan = match self.prepare_operation(request).await {
-                    Ok(operation_plan) => operation_plan,
-                    Err((metadata, response)) => {
-                        let status = response.graphql_status();
-                        sender.send(response).await.ok();
-                        return Err((metadata, status));
-                    }
-                };
-
-                if matches!(operation_plan.ty(), OperationType::Query | OperationType::Mutation) {
-                    let metrics_attributes = Some(operation_plan.metrics_attributes.clone());
-                    let response = self.execute_query_or_mutation(operation_plan).await;
-                    let status = response.graphql_status();
-
-                    sender.send(response).await.ok();
-
-                    Err((metrics_attributes, status))
-                } else {
-                    Ok((self, operation_plan))
-                }
-            })
-            .await;
-
-        let (ctx, operation_plan) = match result {
-            Some(Ok((ctx, operation_plan))) => (ctx, operation_plan),
-            Some(Err((metadata, status))) => return (metadata, status),
-            None => {
-                let response = RequestErrorResponse::gateway_timeout();
-                let status = response.graphql_status();
-                sender.send(response).await.ok();
-                return (None, status);
-            }
-        };
-
-        let mut status: GraphqlResponseStatus = GraphqlResponseStatus::Success;
-
-        struct Sender<'a> {
-            sender: mpsc::Sender<Response>,
-            status: &'a mut GraphqlResponseStatus,
-            operation_name: Option<String>,
-            client: Option<Client>,
-            operation_metrics: &'a GraphqlOperationMetrics,
-        }
-
-        impl crate::execution::ResponseSender for Sender<'_> {
-            type Error = mpsc::SendError;
-            async fn send(&mut self, response: Response) -> Result<(), Self::Error> {
-                *self.status = self.status.union(response.graphql_status());
-
-                let errors = response
-                    .errors()
-                    .iter()
-                    .fold(EnumSet::<ErrorCode>::empty(), |mut set, error| {
-                        set |= error.code;
-                        set
-                    });
-
-                for error in errors {
-                    self.operation_metrics.increment_graphql_errors(GraphqlErrorAttributes {
-                        code: error.into(),
-                        operation_name: self.operation_name.clone(),
-                        client: self.client.clone(),
-                    })
-                }
-
-                self.sender.send(response).await
-            }
-        }
-
-        let metrics_attributes = Some(operation_plan.metrics_attributes.clone());
-
-        ctx.execute_subscription(
-            operation_plan,
-            Sender {
-                sender,
-                status: &mut status,
-                operation_name: metrics_attributes.as_ref().and_then(|a| a.name.clone()),
-                client: client.clone(),
-                operation_metrics: &engine.operation_metrics,
-            },
-        )
-        .await;
-
-        (metrics_attributes, status)
-    }
-
-    async fn prepare_operation(
-        &mut self,
-        request: Request,
-    ) -> Result<ExecutableOperation, (Option<OperationMetricsAttributes>, Response)> {
-        let start = SystemTime::now();
-        let result = self.prepare_operation_inner(request).await;
-        let duration = SystemTime::now().duration_since(start).unwrap_or_default();
-
-        match result {
-            Ok(operation) => {
-                let attributes = QueryPreparationAttributes {
-                    operation_name: operation.prepared.metrics_attributes.name.clone(),
-                    document: Some(operation.prepared.metrics_attributes.sanitized_query.clone()),
-                    success: true,
-                };
-
-                self.operation_metrics.record_preparation_latency(attributes, duration);
-
-                Ok(operation)
-            }
-            Err(e) => {
-                let attributes = QueryPreparationAttributes {
-                    operation_name: None,
-                    document: None,
-                    success: false,
-                };
-
-                self.operation_metrics.record_preparation_latency(attributes, duration);
-
-                Err(e)
-            }
-        }
-    }
-
-    async fn prepare_operation_inner(
-        &mut self,
-        request: Request,
-    ) -> Result<ExecutableOperation, (Option<OperationMetricsAttributes>, Response)> {
-        let result = {
-            let OperationDocument { cache_key, load_fut } = match self.determine_operation_document(&request) {
-                Ok(doc) => doc,
-                // If we have an error a this stage, it means we couldn't determine what document
-                // to load, so we don't consider it a well-formed GraphQL-over-HTTP request.
-                Err(err) => return Err((None, Response::refuse_request_with(http::StatusCode::BAD_REQUEST, err))),
-            };
-
-            if let Some(operation) = self.operation_cache.get(&cache_key).await {
-                self.engine.operation_metrics.record_operation_cache_hit();
-                Ok(operation)
-            } else {
-                self.engine.operation_metrics.record_operation_cache_miss();
-                match load_fut.await {
-                    Ok(document) => Err((cache_key, document)),
-                    Err(err) => return Err((None, Response::request_error([err]))),
-                }
-            }
-        };
-
-        let operation = match result {
-            Ok(operation) => operation,
-            Err((cache_key, document)) => {
-                let operation = Operation::build(&self.schema, &request, &document)
-                    .map(Arc::new)
-                    .map_err(|mut err| (err.take_metrics_attributes(), Response::request_error([err])))?;
-
-                let cache_fut = self.engine.operation_cache.insert(cache_key, operation.clone());
-                self.push_background_future(cache_fut.boxed());
-
-                operation
-            }
-        };
-
-        // GraphQL-over-HTTP spec:
-        //   GET requests MUST NOT be used for executing mutation operations. If the values of {query} and {operationName} indicate that
-        //   a mutation operation is to be executed, the server MUST respond with error status code 405 (Method Not Allowed) and halt
-        //   execution. This restriction is necessary to conform with the long-established semantics of safe methods within HTTP.
-        //
-        // While it's technically a RequestError at this stage, as we have a well-formed GraphQL-over-HTTP request,
-        // we must return a 4xx without regard to the Accept header in all cases, so it's akin to denying the request.
-        //
-        // This error would be confusing for a websocket connection, but today mutation are always
-        // allowed for it.
-        if operation.ty.is_mutation() && !self.request_context.mutations_allowed {
-            return Err((
-                Some(operation.metrics_attributes.clone()),
-                RefusedRequestResponse::method_not_allowed("Mutation is not allowed with a safe method like GET"),
-            ));
-        }
-
-        let variables = Variables::build(self.schema.as_ref(), &operation, request.variables).map_err(|errors| {
-            (
-                Some(operation.metrics_attributes.clone()),
-                Response::request_error(errors),
-            )
-        })?;
-
-        self.finalize_operation(Arc::clone(&operation), variables)
-            .await
-            .map_err(|err| {
-                (
-                    Some(operation.metrics_attributes.clone()),
-                    Response::request_error([err]),
-                )
-            })
-    }
 }
 
 pub struct WebsocketSession<R: Runtime> {
@@ -642,20 +151,11 @@ impl<R: Runtime> Clone for WebsocketSession<R> {
     }
 }
 
-pub(crate) struct RequestContext<C> {
-    pub mutations_allowed: bool,
-    pub headers: http::HeaderMap,
-    pub response_format: ResponseFormat,
-    pub client: Option<Client>,
-    pub access_token: AccessToken,
-    pub hooks_context: C,
-}
-
 impl<R: Runtime> WebsocketSession<R> {
     pub fn execute(&self, event: websocket::SubscribeEvent) -> impl Stream<Item = websocket::Message> {
         let websocket::SubscribeEvent { id, payload } = event;
         self.engine
-            .execute_stream(self.request_context.clone(), payload.0)
+            .execute_websocket_well_formed_graphql_request(self.request_context.clone(), payload.0)
             .map(move |response| match response {
                 Response::RefusedRequest(_) => websocket::Message::Error {
                     id: id.clone(),
@@ -667,17 +167,4 @@ impl<R: Runtime> WebsocketSession<R> {
                 },
             })
     }
-}
-
-fn content_type_is_application_json(headers: &http::HeaderMap) -> bool {
-    static APPLICATION_JSON: http::HeaderValue = http::HeaderValue::from_static("application/json");
-
-    let Some(header) = headers.get(http::header::CONTENT_TYPE) else {
-        return false;
-    };
-
-    let header = header.to_str().unwrap_or_default();
-    let (without_parameters, _) = header.split_once(';').unwrap_or((header, ""));
-
-    without_parameters == APPLICATION_JSON
 }
