+++ engine/crates/engine-v2/src/sources/graphql/record.rs
@@ -17,7 +17,7 @@ pub(super) fn subgraph_retry<R: Runtime>(
     aborted: bool,
 ) {
     ctx.engine
-        .operation_metrics
+        .metrics
         .record_subgraph_retry(SubgraphRequestRetryAttributes {
             name: endpoint.subgraph_name().to_string(),
             aborted,
@@ -31,7 +31,7 @@ pub(super) fn subgraph_duration<R: Runtime>(
     status_code: Option<http::StatusCode>,
     duration: Duration,
 ) {
-    ctx.engine.operation_metrics.record_subgraph_duration(
+    ctx.engine.metrics.record_subgraph_duration(
         SubgraphRequestDurationAttributes {
             name: endpoint.subgraph_name().to_string(),
             subgraph_status,
@@ -46,7 +46,7 @@ pub(super) fn subgraph_request_size<R: Runtime>(
     endpoint: GraphqlEndpointWalker<'_>,
     size: usize,
 ) {
-    ctx.engine.operation_metrics.record_subgraph_request_size(
+    ctx.engine.metrics.record_subgraph_request_size(
         SubgraphRequestBodySizeAttributes {
             name: endpoint.subgraph_name().to_string(),
         },
@@ -59,7 +59,7 @@ pub(super) fn subgraph_response_size<R: Runtime>(
     endpoint: GraphqlEndpointWalker<'_>,
     size: usize,
 ) {
-    ctx.engine.operation_metrics.record_subgraph_response_size(
+    ctx.engine.metrics.record_subgraph_response_size(
         SubgraphResponseBodySizeAttributes {
             name: endpoint.subgraph_name().to_string(),
         },
@@ -72,7 +72,7 @@ pub(super) fn increment_inflight_requests<R: Runtime>(
     endpoint: GraphqlEndpointWalker<'_>,
 ) {
     ctx.engine
-        .operation_metrics
+        .metrics
         .increment_subgraph_inflight_requests(SubgraphInFlightRequestAttributes {
             name: endpoint.subgraph_name().to_string(),
         });
@@ -83,7 +83,7 @@ pub(super) fn decrement_inflight_requests<R: Runtime>(
     endpoint: GraphqlEndpointWalker<'_>,
 ) {
     ctx.engine
-        .operation_metrics
+        .metrics
         .decrement_subgraph_inflight_requests(SubgraphInFlightRequestAttributes {
             name: endpoint.subgraph_name().to_string(),
         });
@@ -91,7 +91,7 @@ pub(super) fn decrement_inflight_requests<R: Runtime>(
 
 pub(super) fn record_subgraph_cache_hit<R: Runtime>(ctx: ExecutionContext<'_, R>, endpoint: GraphqlEndpointWalker<'_>) {
     ctx.engine
-        .operation_metrics
+        .metrics
         .record_subgraph_cache_hit(SubgraphCacheHitAttributes {
             name: endpoint.subgraph_name().to_string(),
         });
@@ -102,7 +102,7 @@ pub(super) fn record_subgraph_cache_miss<R: Runtime>(
     endpoint: GraphqlEndpointWalker<'_>,
 ) {
     ctx.engine
-        .operation_metrics
+        .metrics
         .record_subgraph_cache_miss(SubgraphCacheMissAttributes {
             name: endpoint.subgraph_name().to_string(),
         });
